requirements.txt
absl-py==1.0.0
aiohttp @ file:///D:/bld/aiohttp_1701099640144/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
ansi2html==1.8.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.3.0
astroid==3.0.1
asttokens==2.4.0
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-generator==1.10
async-lru==2.0.4
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1691763562544/work
atomicwrites==1.4.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1683424013410/work
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1698890160476/work
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///D:/bld/cffi_1696001730392/work
chardet==5.2.0
charset-normalizer==2.0.12
click @ file:///D:/bld/click_1692312014553/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cryptography @ file:///D:/bld/cryptography-split_1672672703950/work
cycler==0.12.1
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
dm-tree==0.1.8
entrypoints==0.4
etils==1.5.1
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
Flask==2.2.5
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1685164110706/work
fonttools==4.43.1
fqdn==1.5.1
frozenlist @ file:///D:/bld/frozenlist_1695377787265/work
fsspec==2023.10.0
future==0.18.3
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-auth==2.23.3
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio @ file:///D:/bld/grpc-split_1662408294816/work
h5py==3.10.0
idna==3.3
imageio==2.31.5
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.1.0
infinity==1.5
iniconfig==1.1.1
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_packaging==0.12.3
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
libclang==16.0.6
Markdown==3.3.6
MarkupSafe @ file:///D:/bld/markupsafe_1695367528298/work
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
multidict @ file:///D:/bld/multidict_1696716277911/work
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.7.0
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.0
open3d==0.17.0
opencv-contrib-python==4.8.1.78
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1696448916724/work
overrides==7.4.0
packaging==21.3
paho-mqtt==1.6.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.0.0
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.39
protobuf @ file:///D:/bld/protobuf_1681049992963/work/python
psutil==5.9.6
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments==2.16.1
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1689721553971/work
pylint==3.0.2
pynput==1.7.6
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1685514481738/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///D:/bld/pysocks_1661604991356/work
pytest==7.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rdflib==7.0.0
referencing==0.30.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.8
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tenacity==8.2.3
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1664238338171/work/tensorboard-2.10.1-py3-none-any.whl
tensorboard-data-server @ file:///D:/bld/tensorboard-data-server_1670043904793/work/tensorboard_data_server-0.6.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.10.0
tensorflow-datasets==4.9.2
tensorflow-estimator @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/tensorflow-estimator_1669761460695/work/tensorflow_estimator-2.10.0-py2.py3-none-any.whl
tensorflow-hub==0.12.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
tensorflowjs==3.15.0
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1682317048417/work
terminado==0.17.1
testpath==0.6.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traces==0.6.0
traitlets==5.11.2
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.9
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.2.3
widgetsnbextension==4.0.9
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wrapt==1.15.0
yarl @ file:///D:/bld/yarl_1701168698173/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
